requirements.txt
selenium==4.7.2
python-dotenv==0.21.0
python-telegram-bot==13.7
keyring==23.13.1